Damage prey organs command should now function on all external limbs.  Head removal through this counts as a vore death.

+++ modular_causticcove/code/modules/vore/eating/vore_verbs.dm
@@ -112,14 +112,31 @@
 				visible_message(span_danger("[src] severely damages [T]'s [T_ext.name]!"))
 			else if(B)
 				playsound(T, 'sound/gore/flesh_eat_01.ogg', 100)
-				T_ext.drop_limb(1) //Clean cut so it doesn't kill the prey completely.
-				T_ext.forceMove(B)
-				visible_message(span_warning("[src] swallows [T]'s [T_ext.name] into their [lowertext(B.name)]!"))
+				//Add head check and move organs with limb
+				if (T_ext.body_part == HEAD)
+					T_ext.dismember(BRUTE, BCLASS_CUT, T, T_ext.body_part, 200, FALSE, TRUE)
+					T_ext.forceMove(B)
+					visible_message(span_warning("[src] swallows [T]'s [T_ext.name] into their [lowertext(B.name)]!"))
+					//Set as vore death if head
+					var/mob/dead/observer/G = T.ghostize(TRUE)
+					G.vore_death = TRUE
+				else
+					T_ext.drop_limb(1) //Clean cut so it doesn't kill the prey completely.
+					T_ext.forceMove(B)
+					visible_message(span_warning("[src] swallows [T]'s [T_ext.name] into their [lowertext(B.name)]!"))
 			else
 				playsound(T, 'sound/gore/flesh_eat_01.ogg', 100)
-				T_ext.drop_limb(1) //Clean cut so it doesn't kill the prey completely.
-				T_ext.forceMove(T.loc)
-				visible_message(span_warning("[src] tears off [T]'s [T_ext.name]!"),span_warning("You tear off [T]'s [T_ext.name]!"))
+				if (T_ext.body_part == HEAD)
+					T_ext.dismember(BRUTE, BCLASS_CUT, T, T_ext.body_part, 200, FALSE, TRUE)
+					T_ext.forceMove(B)
+					visible_message(span_warning("[src] tears off [T]'s [T_ext.name]!"),span_warning("You tear off [T]'s [T_ext.name]!"))
+					//Set as vore death if head
+					var/mob/dead/observer/G = T.ghostize(TRUE)
+					G.vore_death = TRUE
+				else
+					T_ext.drop_limb(1) //Clean cut so it doesn't kill the prey completely.
+					T_ext.forceMove(T.loc)
+					visible_message(span_warning("[src] tears off [T]'s [T_ext.name]!"),span_warning("You tear off [T]'s [T_ext.name]!"))
 
 		//Not targeting an internal organ w/ > 25 damage , and the limb doesn't have < 25 damage.
 		else
